{{cookiecutter.project_slug}}/app/requirements/prod.txt
Flask==1.1.2
python-dotenv==0.14.0
flask-sqlalchemy==2.4.4
psycopg2==2.8.5
flask-migrate==2.5.3